daemon: split out /v2/interfaces to api_interfaces_test.go
make it live in daemon_test, also do the same for api_connections.go

they both share the interfacesSuite
