+++ daemon/api_base_test.go
@@ -604,6 +604,10 @@ func handlerCommand(c *check.C, d *Daemon, req *http.Request) (cmd *Command, var
 }
 
 func (s *APIBaseSuite) CheckGetOnly(c *check.C, req *http.Request) {
+	if s.d == nil {
+		panic("call s.Daemon(c) etc in your test first")
+	}
+
 	cmd, _ := handlerCommand(c, s.d, req)
 	c.Check(cmd.POST, check.IsNil)
 	c.Check(cmd.PUT, check.IsNil)
@@ -611,6 +615,10 @@ func (s *APIBaseSuite) CheckGetOnly(c *check.C, req *http.Request) {
 }
 
 func (s *APIBaseSuite) Req(c *check.C, req *http.Request, u *auth.UserState) Response {
+	if s.d == nil {
+		panic("call s.Daemon(c) etc in your test first")
+	}
+
 	cmd, vars := handlerCommand(c, s.d, req)
 	s.vars = vars
 	var f ResponseFunc
@@ -628,4 +636,22 @@ func (s *APIBaseSuite) Req(c *check.C, req *http.Request, u *auth.UserState) Res
 		c.Fatalf("no support for %q for %q", req.Method, req.URL)
 	}
 	return f(cmd, req, u)
+}
+
+func (s *APIBaseSuite) SimulateConflict(name string) {
+	if s.d == nil {
+		panic("call s.Daemon(c) etc in your test first")
+	}
+
+	o := s.d.overlord
+	st := o.State()
+	st.Lock()
+	defer st.Unlock()
+	t := st.NewTask("link-snap", "...")
+	snapsup := &snapstate.SnapSetup{SideInfo: &snap.SideInfo{
+		RealName: name,
+	}}
+	t.Set("snap-setup", snapsup)
+	chg := st.NewChange("manip", "...")
+	chg.AddTask(t)
 }
